# public3/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IZapatillaDB": {
				"properties": {
					"idZapatilla": {
						"type": "number",
						"format": "double"
					},
					"idImagen": {
						"type": "number",
						"format": "double"
					},
					"idModelo": {
						"type": "string"
					},
					"idTalla": {
						"type": "number",
						"format": "double"
					},
					"idFactura": {
						"type": "number",
						"format": "double"
					},
					"precioCompra": {
						"type": "number",
						"format": "double"
					},
					"precioSugerido": {
						"type": "number",
						"format": "double"
					},
					"precioVenta": {
						"type": "number",
						"format": "double"
					},
					"disponibilidad": {
						"type": "number",
						"format": "double"
					},
					"banVendido": {
						"type": "boolean"
					}
				},
				"required": [
					"idZapatilla",
					"idImagen",
					"idModelo",
					"idTalla",
					"idFactura",
					"precioCompra",
					"precioSugerido",
					"precioVenta",
					"disponibilidad",
					"banVendido"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IModeloDB": {
				"properties": {
					"idModelo": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					}
				},
				"required": [
					"idModelo",
					"descripcion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseGeneric": {
				"properties": {
					"payload": {},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"correo"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "pleaser_shoes",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Hector Hernandez"
		}
	},
	"paths": {
		"/mensaje/{nombre}": {
			"get": {
				"operationId": "GetMensaje",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "nombre",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/zapatillas": {
			"get": {
				"operationId": "GetZapatillas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IZapatillaDB"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "SaveZapatillas",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IZapatillaDB"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IZapatillaDB"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateZapatillas",
				"responses": {
					"202": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IZapatillaDB"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IZapatillaDB"
							}
						}
					}
				}
			}
		},
		"/zapatillas/{modelo}": {
			"get": {
				"operationId": "GetZapatillasByModel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IZapatillaDB"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "modelo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/modelos": {
			"get": {
				"operationId": "GetModelo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IModeloDB"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/facturas": {
			"put": {
				"operationId": "GetFacturas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseGeneric"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseGeneric"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}